In the README, clarify that existing fonts are preserved, and reword
One of the things that made me hesitate to install this extension was wondering whether it would wipe my list of fonts. I assuaged my fears by looking inside the extension’s code and then testing it, but it would be better to save people time by explicitly stating how it works.